requirements.txt
aiogram==3.3.0
asyncpg==0.29.0
python-dotenv==1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
tqdm>=4.65.0
chromadb>=0.4.0
ollama>=0.3.0
langchain-core>=0.3.76
PyPDF2>=3.0.0
python-docx>=1.1.0
EbookLib>=0.18
lxml>=4.9.0
requests>=2.31.0
beautifulsoup4>=4.12.3
fake-useragent>=1.4.0
duckduckgo-search>=3.8.6
Telethon==1.41.2
pypdf>=3.9.1
unstructured>=0.12.0
langchain-mistralai>=0.1.8
langchain>=0.1.0
langchain-community>=0.0.33
langchain-core>=0.3.76
langchain-text-splitters>=0.0.1
pandas>=2.0.0
polars>=1.33
datasets>=4.4.0
